splitting of functions
