+++ molecool/atom_data.py
@@ -1,0 +1,26 @@
+"""
+Data used for the rest of the package
+"""
+
+atomic_weights = {
+    'H': 1.00784,
+    'C': 12.0107,
+    'N': 14.0067,
+    'O': 15.999,
+    'P': 30.973762,
+    'F': 18.998403,
+    'Cl': 35.453,
+    'Br': 79.904,
+}
+
+atom_colors = {
+    'H': 'white',
+    'C': '#D3D3D3',
+    'N': '#add8e6',
+    'O': 'red',
+    'P': '#FFA500',
+    'F': '#FFFFE0',
+    'Cl': '#98FB98',
+    'Br': '#F4A460',
+    'S': 'yellow'
+}
